// packages/database/prisma/generated/schemas/objects/ServiceRequestMaxOrderByAggregateInput.schema.ts
import { z } from 'zod';
import { SortOrderSchema } from '../enums/SortOrder.schema';

import type { Prisma } from '../../../../.prisma/client';

const Schema: z.ZodType<Prisma.ServiceRequestMaxOrderByAggregateInput> = z
    .object({
        id: z.lazy(() => SortOrderSchema).optional(),
        type: z.lazy(() => SortOrderSchema).optional(),
        dateCreated: z.lazy(() => SortOrderSchema).optional(),
        dateUpdated: z.lazy(() => SortOrderSchema).optional(),
        status: z.lazy(() => SortOrderSchema).optional(),
        description: z.lazy(() => SortOrderSchema).optional(),
        assignedEmployeeID: z.lazy(() => SortOrderSchema).optional(),
        fromEmployee: z.lazy(() => SortOrderSchema).optional(),
        priority: z.lazy(() => SortOrderSchema).optional(),
    })
    .strict();

export const ServiceRequestMaxOrderByAggregateInputObjectSchema = Schema;